requirements.txt
charset-normalizer==3.4.0
filelock==3.16.1
huggingface-hub==0.26.3
idna==3.10
numpy==2.0.2
packaging==24.2
peft==0.13.2
pillow==11.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
